requirements.txt
fastapi
uvicorn
pydantic
pymupdf
pinecone-client
sentence-transformers
google-generativeai
python-dotenv
